refactor: 미들웨어 디렉토리 구조 변경 / .env 설정 추가 / cross-env 추가
description:
 - .env 파일 수정 요망
 - config.json 파일 수정 요망

 - .env 파일내 설명 작성 했습니다.
