+++ .gitignore
@@ -1,5 +1,15 @@
+# npm
 /node_modules
 package-lock.json
+
+# 환경변수
 .env
+
+# sequelize <--> DB 연결정보
 /config/config.json
+
+# SMTP 
 /config/email.js
+
+# sql문 
+/sql
