+++ controller/Cupload.js
@@ -1,4 +1,3 @@
-const upload = require('../multerConfig'); // Multer 설정 가져오기
 const { User, Question, Board } = require('../models');
 const { Op } = require('sequelize');
 
@@ -59,7 +58,7 @@ exports.uploadEditImageFile = async (req, res) => {
     res.status(200).send({
       success: true,
       msg: '파일이 성공적으로 업로드되었습니다.',
-      url: `http://${URL}/${path}`,
+      url: `${URL}/${path}`,
     });
   } catch (error) {
     console.error('파일 업로드 중 오류 발생:', error);
