+++ middlewares/badWordsFilter/badWordsFilter.js
