+++ models/User.js
@@ -1,5 +1,5 @@
 // 비속어 필터링 모듈 가져오기
-const { filterBadWords, isBadWords } = require('../util/badWordsFilter');
+const { filterBadWords, isBadWords } = require('../middlewares/badWordsFilter/badWordsFilter');
 
 const User = (Sequelize, DataTypes) => {
   const User = Sequelize.define(
@@ -55,9 +55,7 @@ const User = (Sequelize, DataTypes) => {
       hooks: {
         beforeCreate: (record, options) => {
           if (record.dataValues.content) {
-            record.dataValues.content = filterBadWords(
-              record.dataValues.content
-            );
+            record.dataValues.content = filterBadWords(record.dataValues.content);
           }
           if (record.dataValues.title) {
             record.dataValues.title = filterBadWords(record.dataValues.title);
@@ -76,9 +74,7 @@ const User = (Sequelize, DataTypes) => {
 
         beforeBulkUpdate: (options) => {
           if (options.attributes.content) {
-            options.attributes.content = filterBadWords(
-              options.attributes.content
-            );
+            options.attributes.content = filterBadWords(options.attributes.content);
           }
           if (options.attributes.title) {
             options.attributes.title = filterBadWords(options.attributes.title);
