+++ package.json
@@ -1,35 +1,38 @@
 {
-    "name": "sesacproject",
-    "version": "1.0.0",
-    "description": "새싹 팀프로젝트",
-    "main": "app.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "axios": "^1.5.0",
-        "badwords-ko": "^1.0.4",
-        "bcrypt": "^5.1.1",
-        "cookie-parser": "^1.4.6",
-        "dotenv": "^16.3.1",
-        "ejs": "^3.1.9",
-        "express": "^4.18.2",
-        "express-session": "^1.17.3",
-        "multer": "^1.4.4",
-        "mysql": "^2.18.1",
-        "mysql2": "^3.6.0",
-        "nodemailer": "^6.9.5",
-        "path": "^0.12.7",
-        "sanitize-filename": "^1.6.3",
-        "sequelize": "^6.32.1",
-        "sequelize-cli": "^6.6.1",
-        "swagger-jsdoc": "^6.2.8",
-        "swagger-ui-express": "^5.0.0"
-    },
-    "devDependencies": {
-        "nodemon": "^3.0.1"
-    }
+  "name": "sesacproject",
+  "version": "1.0.0",
+  "description": "새싹 팀프로젝트",
+  "main": "app.js",
+  "scripts": {
+    "local": "cross-env NODE_ENV=local DB_ENV=local URL=http://localhost:8000 PORT=8000 nodemon app",
+    "dev": "cross-env NODE_ENV=development DB_ENV=development URL=http://52.79.181.24 PORT=80 node app",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^1.5.0",
+    "badwords-ko": "^1.0.4",
+    "bcrypt": "^5.1.1",
+    "cookie-parser": "^1.4.6",
+    "cross-env": "^7.0.3",
+    "dotenv": "^16.3.1",
+    "ejs": "^3.1.9",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
+    "multer": "^1.4.4",
+    "mysql": "^2.18.1",
+    "mysql2": "^3.6.0",
+    "nodemailer": "^6.9.5",
+    "path": "^0.12.7",
+    "sanitize-filename": "^1.6.3",
+    "sequelize": "^6.32.1",
+    "sequelize-cli": "^6.6.1",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.0"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.1"
+  }
 }
