+++ routes/index.js
@@ -1,13 +1,11 @@
-// (시작) module
-// 경로 선언과 관련된 내용 기술
 const express = require('express');
 const router = express.Router();
+// 컨트롤러
 const Cmain = require('../controller/Cmain');
 const Cuser = require('../controller/Cuser');
 const Cquestion = require('../controller/Cquestion');
 const Cboard = require('../controller/Cboard');
 const Cprofile = require('../controller/Cprofile');
-// const CprofileTest = require("../controller/CprofileTest");
 
 // 메인 페이지 렌더링
 router.get('/', Cmain.getMainPage);
@@ -35,7 +33,6 @@ router.get('/id', Cuser.id);
 
 // 아이디찾기 기능
 router.post('/id', Cuser.findId);
-
 
 // 이메일 인증 페이지 렌더링
